+++ src/team.c
@@ -46,7 +46,7 @@ void team_destroy(struct team *this_team)
 }
 
 
-int team_add_device_sensor(struct team *this_team, unsigned int host_number, char *device_name, char *sensor_name)
+int team_add_device_sensor(struct team *this_team, size_t host_number, char *device_name, char *sensor_name)
 {
     if (this_team != NULL && host_number < this_team->number_of_antennas)
     {
@@ -63,7 +63,7 @@ int team_add_device_sensor(struct team *this_team, unsigned int host_number, cha
 }
 
 
-int team_add_engine_device_sensor(struct team *this_team, unsigned int host_number, char *engine_name, char *device_name, char *sensor_name)
+int team_add_engine_device_sensor(struct team *this_team, size_t host_number, char *engine_name, char *device_name, char *sensor_name)
 {
     if (this_team != NULL && host_number < this_team->number_of_antennas)
     {
